+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Update SaleTunnel to add certificate product course run information.
+- Update SaleTunnel to display product instructions.
 - When a course have multiple seller organizations, CourseGlimpse now display
   the number of seller organization instead of the title of the first one.
 - Update `JOANIE_BACKEND` settings, frontend widgets and learner dashboard